refactor(fan-control): split plugin-storage out of per-game-profiles
Mirror the tdp-control lib layout: the inlined plugin storage helpers
(pluginStoragePath / readPluginStorage / writePluginStorage) move to
lib/plugin-storage.ts and per-game-profiles.ts imports them. Drops the
Bun.write-vs-fs fallback branch (plain node:fs/promises writeFile, same
as tdp-control) and removes the stale comment claiming the fs specifier
matches a mock.module() the tests no longer use. Storage tests split
into lib/plugin-storage.test.ts; the createPluginStoragePersistence
factory tests stay with the engine.
